Changes made after reviews, protocol and classes used.
