+++ Assignment 3/Assignment 3/Node.swift
@@ -1,0 +1,19 @@
+//
+//  Node.swift
+//  Assignment 3
+//
+//  Created by Shubham on 01/02/22.
+//
+
+import Foundation
+
+class Node {
+    let id:Int
+    let nodeName:String
+    let addInfo:String
+    init(id : Int , nodeName: String, addInfo:String){
+        self.id = id
+        self.nodeName = nodeName
+        self.addInfo = addInfo
+    }
+}
